GH-45 : Green mail server test container setup
