+++ src/test/java/com/amalvadkar/ihms/common/AbstractIT.java
@@ -6,6 +6,12 @@ import org.junit.jupiter.api.Tag;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.context.annotation.Import;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.containers.wait.strategy.Wait;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.utility.DockerImageName;
 
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
 
@@ -13,6 +19,18 @@ import static org.springframework.boot.test.context.SpringBootTest.WebEnvironmen
 @Tag("it")
 @Import(ContainersConfig.class)
 public abstract class AbstractIT {
+
+    @Container
+    static GenericContainer<?> greenMailGenericContainer = new GenericContainer(DockerImageName.parse("greenmail/standalone:2.1.0-alpha-4"))
+            .waitingFor(Wait.forLogMessage(".*Starting GreenMail standalone.*", 1))
+            .withEnv("GREENMAIL_OPTS", "-Dgreenmail.setup.test.smtp -Dgreenmail.hostname=0.0.0.0 -Dgreenmail.users=user:admin")
+            .withExposedPorts(3025);
+
+    @DynamicPropertySource
+    static void configureMailHost(DynamicPropertyRegistry registry) {
+        registry.add("spring.mail.host", greenMailGenericContainer::getHost);
+        registry.add("spring.mail.port", greenMailGenericContainer::getFirstMappedPort);
+    }
 
     @LocalServerPort
     int port;
