+++ src/test/java/com/amalvadkar/ihms/common/ContainersConfig.java
@@ -3,13 +3,7 @@ package com.amalvadkar.ihms.common;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
 import org.springframework.context.annotation.Bean;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.containers.MySQLContainer;
-import org.testcontainers.containers.wait.strategy.Wait;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.utility.DockerImageName;
 
 @TestConfiguration(proxyBeanMethods = false)
 public class ContainersConfig {
@@ -18,18 +12,6 @@ public class ContainersConfig {
     @ServiceConnection
     MySQLContainer<?> mysqlContainer(){
         return new MySQLContainer<>("mysql:8.0-debian");
-    }
-
-    @Container
-    static GenericContainer<?> greenMailGenericContainer = new GenericContainer(DockerImageName.parse("greenmail/standalone:2.1.0-alpha-4"))
-            .waitingFor(Wait.forLogMessage(".*Starting GreenMail standalone.*", 1))
-            .withEnv("GREENMAIL_OPTS", "-Dgreenmail.setup.test.smtp -Dgreenmail.hostname=0.0.0.0 -Dgreenmail.users=user:admin")
-            .withExposedPorts(3025);
-
-    @DynamicPropertySource
-    static void configureMailHost(DynamicPropertyRegistry registry) {
-        registry.add("spring.mail.host", greenMailGenericContainer::getHost);
-        registry.add("spring.mail.port", greenMailGenericContainer::getFirstMappedPort);
     }
 
 }
